WanderVerse/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cfgv==3.4.0
distlib==0.3.8
Django==5.1.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
filelock==3.13.1
firebase-admin==6.7.0
gunicorn==21.2.0
identify==2.5.34
markdown==3.6
nodeenv==1.8.0
packaging==23.1
Pillow==10.3.0
pip==24.0
platformdirs==4.2.0
pre-commit==3.6.1
psycopg2-binary==2.9.9
python-dotenv==1.1.0
python-magic==0.4.27
PyYAML==6.0.1
sentry-sdk==1.40.5
setuptools==68.2.2
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.2.1
virtualenv==20.25.0
wheel==0.41.2
requests==2.31.0
django-tailwind==3.8.0
django-compressor==4.4
django-browser-reload==1.12.1
django-cors-headers==4.7.0
overpy==0.5